Implement more nuanced ballot parsing.
Rules can now be set for minimum and maximum numbers of votes above and below the line.

As a side-effect, the mapping of preference <-> candidate ID was reversed in the `PrefMap`
type, which lead to the discovery that loads of NSW BTL votes from 2013 are actually invalid due
to multiple candidates/preference! Weird!

Looks like ballot parsing and validation is the most finnicky aspect of running an election.

Closes #4.